add useFirebaseUi
